feat: customHook to articleDashboard
